delete duplicated code
